% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute.contrasts}
\alias{compute.contrasts}
\title{Computes contrasts}
\usage{
compute.contrasts(x, y, type = c("Fisher", "simple"))
}
\arguments{
\item{x}{n by p design matrix}

\item{y}{binary (0 or 1) vector of length n indicating class}

\item{type}{determines whether Fisher transform should be applied to interaction contrasts.  Default: Fisher}
}
\value{
Returns w and z.
}
\description{
Computes the main effect and interaction contrasts, "w" and "z", respectively.
}

