% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{hiertest-package}
\alias{hiertest-package}
\title{Convex Hierarchical Testing of Interactions.}
\description{
Implementation of the convex hierarchical testing (CHT) procedure introduced in
Bien, Simon, and Tibshirani (2015) Convex Hierarchical Testing of
Interactions. Annals of Applied Statistics. Vol. 9, No. 1, 27-42.
Its main functions are \code{\link{hiertest}} and \code{\link{estimate.fdr}}.
}
\author{
Jacob Bien, Noah Simon, and Rob Tibshirani
}
\references{
Bien, Simon, and Tibshirani (2015) Convex Hierarchical Testing
of Interactions. Annals of Applied Statistics. Vol. 9, No. 1, 27-42.
}

