% Generated by roxygen2 (4.0.2): do not edit by hand
\name{interactions.above}
\alias{interactions.above}
\title{Get significant interactions above a threshold}
\usage{
interactions.above(testobj, lambda)
}
\arguments{
\item{testobj}{output of \link{hiertest}}

\item{lambda}{a threshold above which we're interested in interactions}
}
\description{
Returns all interactions whose statistics are above a certain level
}

