% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.estfdr}
\alias{plot.estfdr}
\title{Plot the estimated FDR}
\usage{
\method{plot}{estfdr}(x, ...)
}
\arguments{
\item{x}{output of \link{estimate.fdr}}

\item{...}{additional arguments to pass to plot.}
}
\description{
Plot the estimated FDR
}

