% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.hiertest}
\alias{plot.hiertest}
\title{Plot Convex Hierarchical Testing Statistics}
\usage{
\method{plot}{hiertest}(x, ...)
}
\arguments{
\item{x}{output of \link{hiertest}.}

\item{...}{additional arguments to \code{plot}}
}
\description{
This plots the main effect and interaction statistics from the CHT procedure.
}

