% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.estfdr}
\alias{print.estfdr}
\title{Print the estimated FDR}
\usage{
\method{print}{estfdr}(x, ...)
}
\arguments{
\item{x}{output of \link{estimate.fdr}}

\item{...}{additional arguments to pass to print.data.frame.}
}
\description{
Print the estimated FDR
}

