% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.hiertest}
\alias{print.hiertest}
\title{Print Convex Hierarchical Testing Statistics}
\usage{
\method{print}{hiertest}(x, nshow = 20, ...)
}
\arguments{
\item{x}{output of \link{hiertest}}

\item{nshow}{show the largest \code{nshow} statistics}

\item{...}{additional arguments to \code{print.data.frame}}
}
\description{
This prints the top main effect and interaction statistics from the CHT
procedure. Use \link{estimate.fdr} to get an appropriate cutoff.
}

